+++ nengo/tests/test_pytest.py
@@ -1,3 +1,5 @@
+import pytest
+
 import nengo.conftest
 import nengo.utils.numpy as npext
 
@@ -10,31 +12,74 @@ def test_seed_fixture(seed):
     assert i == 1832276344
 
 
-def test_unsupported(testdir):
-    testdir.makeconftest(
-        """
-        import nengo.conftest
+@pytest.mark.parametrize("use_ini", (True, False))
+@pytest.mark.parametrize("xfail", (True, False))
+def test_unsupported(use_ini, xfail, testdir):
+    testdir.makefile(".py", test_file="""
+        import pytest
 
-        class MockSimulator(object):
-            unsupported = [('*', 'mock simulator')]
+        @pytest.mark.parametrize("param", (True, False))
+        def test_unsupported(param):
+            assert param
 
-        nengo.conftest.TestConfig.Simulator = MockSimulator
+        @pytest.mark.parametrize("param", (True, False))
+        def test_unsupported_all(param):
+            assert False
+
+        def test_supported():
+            assert True
         """)
-    outcomes = testdir.runpytest(
-        "-p", "nengo.tests.options", "--pyargs", "nengo",
-    ).parseoutcomes()
-    assert outcomes["skipped"] > 350
-    assert outcomes["deselected"] > 700
-    assert "passed" not in outcomes
-    assert "failed" not in outcomes
 
-    outcomes = testdir.runpytest(
-        "-p", "nengo.tests.options", "--pyargs", "nengo", "--unsupported",
-    ).parseoutcomes()
-    assert outcomes["xfailed"] > 350
-    assert outcomes["deselected"] > 700
-    assert "passed" not in outcomes
-    assert "failed" not in outcomes
+    if use_ini:
+        testdir.makefile(".ini", pytest="""
+            [pytest]
+            nengo_test_unsupported =
+                test_file.py:test_unsupported[False]
+                    "One unsupported param
+                    with multiline comment"
+                test_file.py:test_unsupported_all*
+                    "Two unsupported params
+                    with multiline comment"
+            """)
 
-    # runpytest runs in-process, so we have to undo changes to TestConfig
-    nengo.conftest.TestConfig.Simulator = nengo.Simulator
+        testdir.makefile(".py", conftest="""
+            from nengo.conftest import pytest_runtest_setup, pytest_configure
+            """)
+    else:
+        testdir.makefile(".py", conftest=
+            """
+            from nengo.conftest import (
+                TestConfig, pytest_runtest_setup, pytest_configure)
+    
+            class MockSimulator(object):
+                unsupported = [
+                    ('test_file.py:test_unsupported[False]', 
+                     'One unsupported param '
+                     'with multiline comment'),
+                    ('test_file.py:test_unsupported_all*',
+                     'Two unsupported params '
+                     'with multiline comment'),
+                ]
+    
+            TestConfig.Simulator = MockSimulator
+            """)
+
+    args = "-p nengo.tests.options -rsx".split()
+    if xfail:
+        args.append("--unsupported")
+    output = testdir.runpytest(*args)
+
+    output.stdout.fnmatch_lines_random([
+        "*One unsupported param with multiline comment",
+        "*Two unsupported params with multiline comment",
+    ])
+
+    outcomes = output.parseoutcomes()
+    if xfail:
+        assert outcomes["xfailed"] == 3
+        assert "skipped" not in outcomes
+    else:
+        assert outcomes["skipped"] == 3
+        assert "xfailed" not in outcomes
+    assert "failed" not in outcomes
+    assert outcomes["passed"] == 2